add separate input file

+++ turbulent_main.py
@@ -1,3 +1,4 @@
+from input_file import *
 from  olfactory_lib import *
 import pandas as pd
 import multiprocessing as mp
@@ -31,112 +32,7 @@ def parallel_run(n):
 
     return arrival_time, agents_in_Rb, success, seed
 
-# TODO: extend the simulation box (pad height)
-# move the arrow on the CoM
-# launch simulations with shift and try to reobtain the previous results (change N?)J
-
-# plotting parameters 
-real_time_plot = True
-plot_flow = False
-save_frames = True
-pause_time = 0.001
-
-# read h5 flow file or local npy file
-read_h5 = True
-
-# visual_radius = 0.1 # Ra
-# visual_radius = 0.5 # Ra
-# visual_radius = 1 # Ra
-# visual_radius = 5 # Ra
-visual_radius = 10 # Ra
-
-# vertical shift of the initial position (in perc of height/2)
-shift = 0.4
-
-reach_radius = 0.4
-
-# time parameters
-decision_time = 1 # Δt
-
-# smelling threshold
-threshold = 0.0008
-
-# elastic constant
-kelast = 1
-
-# number of agents
-n_agents = 100 # N
-
-# do more runs at the same time
-parallel = True
-n_threads = 6 # number of threads used for parallelisation
- 
-# number of successful episodes to sample
-n_samples = 50
-
-# max number simulations to run to reach the sampling limit
-limit = int(n_samples*10)
-
-# use elastic recall force
-elastic = True
-
-# use a stochastic or a turbulent flow
-turbulent = True
-
-# use a different beta for informed and uninformed agents
-adaptive_beta = False
-
-# path of the turbulent flow
-if read_h5: path = '/storage/boccardo/odor_data_re280_small_source/r280_small_source.h5'
-else: path = 'flow/re280_small_source'
-
 set_h5_flag(read_h5)
-
-# trust parameter (β) values to check in a parallel run
-trusts = np.round(np.arange(0.0, 1.1, 0.1),2) 
-
-# constant trust parameter
-# trust = 0.85 # β
-trust = 0.05 # β
-
-# these are relevant only if we are using an adaptive beta
-trust_uninform = 0.9
-trust_inform = 0.1
-
-# decay time used both for beta and for the surging phase
-decay_time = 8
-
-Rd = 0.4 # olfactory range
-Lx = 50 # distance from the source
-
-# length of the simulation box (height is given by the flow data)
-length = 100
-
-# parameters of the agents
-speed = 0.2 # v0
-olfactory_radius = Rd # Rd 
-memory_time = 1/decision_time # inverse of λ
-sensing_noise = 0.1 # eta
-spawn_radius = 5 # Rb
-
-# parameters of the particle cloud
-particle_dt = decision_time/10 # δt
-particle_rate = 10 # J
-flow_dt = particle_dt
-
-# parameters of the stochastic flow
-fluct_intensity = 0.42
-flow_lengthscale = 10
-flow_corr_time = 5
-mean_wind = [1, 0]
-loop_cycles = 10
-
-# max duration of the simulation
-Ts = Lx/speed # straight-path time
-final_time = 10*Ts 
-
-# name of the output results file
-filename = f'r280_ra{visual_radius}_dt{decision_time}_thr{threshold}_k{kelast}_shift{shift}_N{n_agents}'
 
 # check if file already exists
 if parallel and os.path.isfile(f'results/{filename}.pkl'):
@@ -159,7 +55,7 @@ flow = Flow_turbulent(path, length)
 cloud = Cloud_turbulent(flow)
 
 # delete old frames
-if save_frames:
+if save_frames and not parallel:
     import os
     os.system(f"rm -f frames/frame*.png")
